requirements.txt
Flask==2.0
mysql-connector-python
